Split map into its own file
