+++ globals.py
@@ -1,0 +1,6 @@
+WIDTH = 1920
+HEIGHT = 1080
+MARGIN = 800
+PADDING = 25
+COL_PART = 10
+ROW_PART = 5
